feature: 添加草稿箱功能
现在提交文档的时候可以选择是否保存为草稿。草稿的标签不会被记录，只有当草稿被发布后它的标签才会起作用。
变更：
- api-define.ts 添加了草稿功能相关的api
- DocList.vue 添加了新的props，允许父组件指定单行的点击事件，同时调整了样式代码
- Milkdown.vue 添加了对草稿功能的支持
- MiniOptionsPanel.vue 添加了我的草稿页面的链接
- MyDrafts.vue 我的草稿
- main.ts 添加了草稿相关的路由
- theme.ts 添加了调整配色透明度的函数

+++ src/components/Milkdown.vue
@@ -1,39 +1,61 @@
 <template>
-  <div :class="{'components-root':true,
-      'editor-to-tight': (!outLineFirstLoaded)&&(showOutline),
-      'editor-to-wide': (!outLineFirstLoaded)&&(!showOutline)}">
-    <VueEditor id="milkdown-root" :editor="editor"/>
-    <div class="side-icon-bar"
-         :style="{ backgroundColor: customComponentThemeProvider.colorSet.halfDeep}">
-      <n-icon :color="colorSet.fontColor2" size="24" class="icon bar-controller" @click="showSideBar=!showSideBar">
-        <ChevronDown48Regular/>
+  <div :class="{
+    'components-root': true,
+    'editor-to-tight': (!outLineFirstLoaded) && (showOutline),
+    'editor-to-wide': (!outLineFirstLoaded) && (!showOutline)
+  }">
+  <n-spin class="uploading-mask" v-show="uploading"></n-spin>
+    <VueEditor id="milkdown-root" :editor="editor" />
+    <div class="side-icon-bar" :style="{ backgroundColor: customComponentThemeProvider.colorSet.halfDeep }">
+      <n-icon :color="colorSet.fontColor2" size="24" class="icon bar-controller" @click="showSideBar = !showSideBar">
+        <ChevronDown48Regular />
       </n-icon>
-      <div :class="{'side-icon-container':true,
-      'show-bar': (!sideBarFirstLoaded)&&(showSideBar),
-      'hide-bar': (!sideBarFirstLoaded)&&(!showSideBar)}">
-        <n-icon :color="colorSet.fontColor2" size="24" class="icon upload-icon" @click="showOutline=!showOutline">
-          <TextBulletListTree16Regular/>
+      <div :class="{
+        'side-icon-container': true,
+        'show-bar': (!sideBarFirstLoaded) && (showSideBar),
+        'hide-bar': (!sideBarFirstLoaded) && (!showSideBar)
+      }">
+        <n-icon :color="colorSet.fontColor2" size="24" class="icon" @click="showOutline = !showOutline">
+          <TextBulletListTree16Regular />
         </n-icon>
-        <n-icon :color="colorSet.fontColor2" size="24" class="icon upload-icon" @click="fullScreenStatus=!fullScreenStatus">
+        <n-icon :color="colorSet.fontColor2" size="24" class="icon" @click="fullScreenStatus = !fullScreenStatus">
           <FullScreenMaximize24Regular v-show="!fullScreenStatus" />
           <FullScreenMinimize24Regular v-show="fullScreenStatus" />
         </n-icon>
         <div v-show=!readOnly>
-          <n-icon :color="colorSet.fontColor2" size="24" class="icon upload-icon" @click="uploadDoc">
-            <CloudUploadOutlined/>
+          <n-icon :color="colorSet.fontColor2" size="24" class="icon" @click="showPictureSelectModal = true">
+            <n-tooltip trigger="hover" placement="left-start">
+              <template #trigger>
+                <PictureOutlined />
+              </template>
+              上传图片
+            </n-tooltip>
           </n-icon>
-          <n-icon :color="colorSet.fontColor2" size="24" class="icon create-picture-url-icon"
-                  @click="showPictureSelectModal=true">
-            <PictureOutlined/>
+          <n-icon :color="colorSet.fontColor2" size="24" class="icon" @click="uploadDoc(false)">
+            <n-tooltip trigger="hover" placement="left-start">
+              <template #trigger>
+                <CloudUploadOutlined />
+              </template>
+              {{ iconNotification('uploadDoc') }}
+            </n-tooltip>
+          </n-icon>
+          <n-icon :color="colorSet.fontColor2" size="24" class="icon" @click="uploadDoc(true)"
+            v-show="router.currentRoute.value.name !== 'editDoc'">
+            <n-tooltip trigger="hover" placement="left-start">
+              <template #trigger>
+                <Save16Regular />
+              </template>
+              {{ iconNotification('uploadDraft') }}
+            </n-tooltip>
           </n-icon>
         </div>
       </div>
     </div>
     <n-modal v-model:show="showPictureSelectModal" display-directive="show">
       <PictureUpload :extend-actions="imageActions"
-                     style="position: absolute;top:50%;left:50%;transform: translate(-50%,-50%)"/>
+        style="position: absolute;top:50%;left:50%;transform: translate(-50%,-50%)" />
     </n-modal>
-    <DocOutline :outline="docOutlines" :outline-show="showOutline"/>
+    <DocOutline :outline="docOutlines" :outline-show="showOutline" />
   </div>
 </template>
 <script lang="ts">
@@ -48,38 +70,38 @@ export default {
 </script>
 
 <script setup lang="ts">
-import {defaultValueCtx, Editor, editorViewOptionsCtx, rootCtx, editorCtx} from '@milkdown/core';
-import {clipboard} from '@milkdown/plugin-clipboard';
-import {cursor} from '@milkdown/plugin-cursor';
-import {diagram} from '@milkdown/plugin-diagram';
-import {emoji} from '@milkdown/plugin-emoji';
-import {history} from '@milkdown/plugin-history';
-import {indent} from '@milkdown/plugin-indent';
-import {listener, listenerCtx} from '@milkdown/plugin-listener';
-import {math} from '@milkdown/plugin-math';
-import {menu} from '@milkdown/plugin-menu';
-import {prism} from '@milkdown/plugin-prism';
-import {slash} from '@milkdown/plugin-slash';
-import {tooltip} from '@milkdown/plugin-tooltip';
-import {upload} from '@milkdown/plugin-upload';
-import {gfm} from '@milkdown/preset-gfm';
-import {nordDark, nordLight} from '@milkdown/theme-nord';
-import {insert, replaceAll, forceUpdate, switchTheme,outline} from '@milkdown/utils';
-import {VueEditor, useEditor} from "@milkdown/vue";
-import {defineComponent, defineProps, ref, Ref, onMounted, watch, computed} from "vue";
-import {CloudUploadOutlined, PictureOutlined} from "@vicons/antd";
-import {ChevronDown48Regular, TextBulletListTree16Regular,FullScreenMaximize24Regular,FullScreenMinimize24Regular} from '@vicons/fluent'
-import {NIcon, NEllipsis, NDialogProvider, useDialog, useNotification, NModal, NCard} from "naive-ui";
+import { defaultValueCtx, Editor, editorViewOptionsCtx, rootCtx, editorCtx } from '@milkdown/core';
+import { clipboard } from '@milkdown/plugin-clipboard';
+import { cursor } from '@milkdown/plugin-cursor';
+import { diagram } from '@milkdown/plugin-diagram';
+import { emoji } from '@milkdown/plugin-emoji';
+import { history } from '@milkdown/plugin-history';
+import { indent } from '@milkdown/plugin-indent';
+import { listener, listenerCtx } from '@milkdown/plugin-listener';
+import { math } from '@milkdown/plugin-math';
+import { menu } from '@milkdown/plugin-menu';
+import { prism } from '@milkdown/plugin-prism';
+import { slash } from '@milkdown/plugin-slash';
+import { tooltip } from '@milkdown/plugin-tooltip';
+import { upload } from '@milkdown/plugin-upload';
+import { gfm } from '@milkdown/preset-gfm';
+import { nordDark, nordLight } from '@milkdown/theme-nord';
+import { insert, replaceAll, forceUpdate, switchTheme, outline } from '@milkdown/utils';
+import { VueEditor, useEditor } from "@milkdown/vue";
+import { defineComponent, defineProps, ref, Ref, onMounted, watch, computed } from "vue";
+import { CloudUploadOutlined, PictureOutlined } from "@vicons/antd";
+import { ChevronDown48Regular, TextBulletListTree16Regular, FullScreenMaximize24Regular, FullScreenMinimize24Regular, Save16Regular } from '@vicons/fluent'
+import { NIcon, NEllipsis, NDialogProvider, useDialog, useNotification, NModal, NCard, NTooltip, NSpin } from "naive-ui";
 import axios from 'axios';
-import {AxiosRequestConfig} from 'axios'
-import {DocApi, TagApi} from '../api-define'
-import {vue} from '../main'
-import {useRoute, useRouter, onBeforeRouteLeave, onBeforeRouteUpdate} from 'vue-router'
-import {loginStatus,fullScreenStatus} from '../globalStatus'
+import { AxiosRequestConfig } from 'axios'
+import { DocApi, TagApi } from '../api-define'
+import { vue } from '../main'
+import { useRoute, useRouter, onBeforeRouteLeave, onBeforeRouteUpdate } from 'vue-router'
+import { loginStatus, fullScreenStatus } from '../globalStatus'
 import PictureUpload from './PictureUpload.vue'
-import {ImageAction} from '../model/models'
-import {blankTransport} from '../tooles'
-import {customComponentThemeProvider} from '../theme'
+import { ImageAction } from '../model/models'
+import { blankTransport } from '../tooles'
+import { customComponentThemeProvider,getOpacityColor } from '../theme'
 import DocOutline from './sub/DocOutline.vue'
 
 /**
@@ -128,44 +150,44 @@ const editorRef: Ref<Editor | null> = ref(null);
 /**
  * 创建编辑器
  */
-const {editor} = useEditor((root) =>
-        Editor.make()
-            .config((ctx) => {
-              ctx.set(rootCtx, root);
-              //监听文档内容更改
-              ctx.get(listenerCtx)
-                  .mounted((ctx) => {
-                    //获取并绑定编辑器对象
-                    editorRef.value = ctx.get(editorCtx);
-                    loadDoc(ctx.get(editorCtx));
-                    loginStatus.registerAction(() => loadDoc(ctx.get(editorCtx)));
-                  })
-                  .markdownUpdated((ctx, markdown, prevMarkdown) => {
-                    refreshOutline();
-                    content.value = markdown;
-                  });
-              //设置只读
-              ctx.set(editorViewOptionsCtx, {editable});
-              //设置默认值
-              ctx.set(defaultValueCtx, '');
-            })
-            //.use(nordDark)
-            .use(customComponentThemeProvider.value.editorTheme)
-            .use(gfm)
-            .use(listener)
-            .use(clipboard)
-            .use(history)
-            .use(cursor)
-            .use(prism)
-            .use(diagram)
-            .use(tooltip)
-            .use(math)
-            .use(emoji)
-            .use(indent)
-            .use(upload)
-            .use(slash)
-            .use(menu)
-    ,)
+const { editor } = useEditor((root) =>
+  Editor.make()
+    .config((ctx) => {
+      ctx.set(rootCtx, root);
+      //监听文档内容更改
+      ctx.get(listenerCtx)
+        .mounted((ctx) => {
+          //获取并绑定编辑器对象
+          editorRef.value = ctx.get(editorCtx);
+          loadDoc(ctx.get(editorCtx));
+          loginStatus.registerAction(() => loadDoc(ctx.get(editorCtx)));
+        })
+        .markdownUpdated((ctx, markdown, prevMarkdown) => {
+          refreshOutline();
+          content.value = markdown;
+        });
+      //设置只读
+      ctx.set(editorViewOptionsCtx, { editable });
+      //设置默认值
+      ctx.set(defaultValueCtx, '');
+    })
+    //.use(nordDark)
+    .use(customComponentThemeProvider.value.editorTheme)
+    .use(gfm)
+    .use(listener)
+    .use(clipboard)
+    .use(history)
+    .use(cursor)
+    .use(prism)
+    .use(diagram)
+    .use(tooltip)
+    .use(math)
+    .use(emoji)
+    .use(indent)
+    .use(upload)
+    .use(slash)
+    .use(menu)
+  ,)
 
 //尝试获取url中的docId
 const docIdInPath = useRoute().params.docId;
@@ -257,10 +279,10 @@ function loadMetadata(content: string) {
 /**
  * 根据情况切换不同的api
  */
-function getUploadAPI(): AxiosRequestConfig[] {
+function getUploadAPI(isDraft: boolean): AxiosRequestConfig[] {
   const result: AxiosRequestConfig[] = [];
-  //新增文档的情况
-  if (router.currentRoute.value.name === 'newDoc') {
+  //新增文档
+  if (router.currentRoute.value.name === 'newDoc' && !isDraft) {
     if (tags.value.length === 0) {
       result.push(DocApi.newDoc(title.value, content.value));
     } else {
@@ -268,19 +290,51 @@ function getUploadAPI(): AxiosRequestConfig[] {
     }
     return result;
   }
-  //修改旧文档的情况
-  const docId = getDocId();
-  result.push(DocApi.changeDocName(title.value, docId));
-  result.push(DocApi.modifyDoc(docId, content.value));
-  if (tags.value.length === 0) {
-    result.push(TagApi.deleteDocMap(docId));
-  } else {
-    result.push(TagApi.setMap(docId, tags.value));
+  //新增草稿
+  if (router.currentRoute.value.name === 'newDoc' && isDraft) {
+    result.push(DocApi.saveDocAsDraft(title.value, content.value));
+    return result;
+  }
+  //修改文档
+  if (router.currentRoute.value.name === 'editDoc' && !isDraft) {
+    const docId = getDocId();
+    result.push(DocApi.changeDocName(title.value, docId));
+    result.push(DocApi.modifyDoc(docId, content.value));
+    if (tags.value.length === 0) {
+      result.push(TagApi.deleteDocMap(docId));
+    } else {
+      result.push(TagApi.setMap(docId, tags.value));
+    }
+    return result;
+  }
+  if (router.currentRoute.value.name === 'editDoc' && isDraft) {
+    notification.create({
+      title: '错误',
+      duration: 2500,
+      content: '已发布的文档不能作为草稿保存',
+      closable: false
+    })
+    return result;
+  }
+  //修改草稿
+  if (router.currentRoute.value.name === 'editDraft' && isDraft) {
+    const docId = getDocId();
+    result.push(DocApi.changeDocName(title.value, docId));
+    result.push(DocApi.modifyDoc(docId, content.value));
+    return result;
+  }
+  //发布草稿
+  if (router.currentRoute.value.name === 'editDraft' && !isDraft) {
+    const docId = getDocId();
+    result.push(DocApi.changeDocName(title.value, docId));
+    result.push(DocApi.modifyDoc(docId, content.value));
+    result.push(DocApi.submitDraft(docId, tags.value));
+    return result;
   }
   return result;
 }
 
-
+//文档提交监控
 let apiFailedCount = 0;
 const apiCount: Ref<number> = ref(0);
 let failedMessage = ''
@@ -304,14 +358,14 @@ watch(apiCount, async (newCount, oldCount) => {
   }
 })
 
+const uploading: Ref<boolean> = ref(false);
+
 /**
  * 上传文档
  */
-function uploadDoc() {
+async function uploadDoc(isDraft: boolean = false) {
+  uploading.value = true;
   loadMetadata(content.value);
-  /*  console.log('content:\n'+content.value);
-    console.log('title:\n'+title.value);
-    console.log('tags:\n'+tags.value);*/
   //判断是否可提交
   if (title.value === '') {
     dialog.warning({
@@ -319,39 +373,64 @@ function uploadDoc() {
       content: '文档缺少标题，提交失败',
       positiveText: '确定'
     })
+    uploading.value = false;
     return;
   }
   //上传
-  const apis: AxiosRequestConfig[] = getUploadAPI();
+  const apis: AxiosRequestConfig[] = getUploadAPI(isDraft);
+  if (apis.length === 0) {
+    uploading.value = false;
+    return;
+  }
   failedMessage = '';
   apiCount.value = apis.length;//重置
   apiFailedCount = 0;//重置清零
   for (let i in apis) {
-    axios.request(apis[i]).then(
-        (response) => {
-          if (response.data.code !== '00000') {
-            console.log(response.data);
-            if (failedMessage === '') {
-              failedMessage = response.data.message;
-            }
-            apiFailedCount++;
-          }
-          if (response.data.code === 'A0200') {
-            //loginStatus.loginFailed();
-          }
-          apiCount.value--;
-        }
-    ).catch(
-        (response) => {
-          console.log(response.data)
-          apiFailedCount++;
-          apiCount.value--;
-          if (failedMessage === '' && response.data !== undefined && response.data.message !== undefined) {
-            failedMessage = response.data.message;
-          }
-        }
-    )
+    //   axios.request(apis[i]).then(
+    //       (response) => {
+    //         if (response.data.code !== '00000') {
+    //           console.log(response.data);
+    //           if (failedMessage === '') {
+    //             failedMessage = response.data.message;
+    //           }
+    //           apiFailedCount++;
+    //         }
+    //         if (response.data.code === 'A0200') {
+    //           //loginStatus.loginFailed();
+    //         }
+    //         apiCount.value--;
+    //       }
+    //   ).catch(
+    //       (response) => {
+    //         console.log(response.data)
+    //         apiFailedCount++;
+    //         apiCount.value--;
+    //         if (failedMessage === '' && response.data !== undefined && response.data.message !== undefined) {
+    //           failedMessage = response.data.message;
+    //         }
+    //       }
+    //   )
+    // }
+    try {
+      const response = await axios.request(apis[i]);
+      if (response.data.code !== '00000') {
+        failedMessage = response.data.message;
+        apiFailedCount++;
+        apiCount.value = 0;
+        uploading.value = false;
+        return;
+      }
+      apiCount.value--;
+    } catch (err) {
+      failedMessage = '请求异常，请联系系统管理员';
+      apiFailedCount++;
+      apiCount.value = 0;
+      uploading.value = false;
+      return;
+    }
   }
+  uploading.value = false;
+  router.push({ name: 'myDoc' });
 }
 
 const showPictureSelectModal: Ref<boolean> = ref(false);
@@ -390,13 +469,36 @@ const outLineFirstLoaded: Ref<boolean> = ref(true);
 watch(showOutline, async (newV, oldV) => {
   outLineFirstLoaded.value = false;
 })
-const docOutlines:Ref<any> = ref([]);
-function refreshOutline(){
+const docOutlines: Ref<any> = ref([]);
+function refreshOutline() {
   docOutlines.value = editorRef.value?.action(outline());
 }
 
+//按钮提示内容控制
+function iconNotification(iconName: string): string {
+  if (router.currentRoute.value.name === 'newDoc' && iconName === 'uploadDoc') {
+    return '上传文档'
+  };
+  if (router.currentRoute.value.name === 'newDoc' && iconName === 'uploadDraft') {
+    return '保存文档为草稿'
+  };
+  if (router.currentRoute.value.name === 'editDoc' && iconName === 'uploadDoc') {
+    return '更新文档'
+  };
+  if (router.currentRoute.value.name === 'editDoc' && iconName === 'uploadDraft') {
+    return '不支持的操作'
+  };
+  if (router.currentRoute.value.name === 'editDraft' && iconName === 'uploadDoc') {
+    return '发布草稿'
+  };
+  if (router.currentRoute.value.name === 'editDraft' && iconName === 'uploadDraft') {
+    return '保存草稿'
+  };
+  return '错误';
+}
+
 //离开编辑器后自动退出编辑器全屏
-onBeforeRouteLeave((to,from)=>{
+onBeforeRouteLeave((to, from) => {
   fullScreenStatus.value = false;
 });
 
@@ -423,6 +525,7 @@ onBeforeRouteLeave((to,from)=>{
   0% {
     width: 90%
   }
+
   100% {
     width: calc(90% - 200px)
   }
@@ -432,6 +535,7 @@ onBeforeRouteLeave((to,from)=>{
   0% {
     width: calc(90% - 200px)
   }
+
   100% {
     width: 90%
   }
@@ -455,12 +559,12 @@ onBeforeRouteLeave((to,from)=>{
   display: none;
 }
 
-#milkdown-root > :deep(.milkdown) {
+#milkdown-root> :deep(.milkdown) {
   height: 100%;
   overflow-y: scroll;
 }
 
-#milkdown-root > :deep(.milkdown::-webkit-scrollbar) {
+#milkdown-root> :deep(.milkdown::-webkit-scrollbar) {
   display: none;
 }
 
@@ -508,15 +612,17 @@ onBeforeRouteLeave((to,from)=>{
   0% {
     height: 0px;
   }
+
   100% {
-    height: 216px;
+    height: 270px;
   }
 }
 
 @keyframes bar-container-hide {
   0% {
-    height: 216px;
+    height: 270px;
   }
+
   100% {
     height: 0px;
   }
@@ -526,8 +632,19 @@ onBeforeRouteLeave((to,from)=>{
   margin: 15px 15px 15px 15px;
 }
 
+.icon :focus {
+  outline-width: 0px;
+}
+
 .side-icon-container .icon:hover {
   cursor: pointer;
 }
 
+.uploading-mask{
+  position:absolute;
+  height: 100%;
+  width: 100%;
+  z-index: 1;
+  background-color:v-bind(getOpacityColor(colorSet.deep,0.7));
+}
 </style>
